stunpaper cult

+++ code/game/objects/structures/plasticflaps.dm
@@ -1,93 +1,28 @@
-/obj/structure/plasticflaps
-	name = "airtight plastic flaps"
-	desc = "Heavy duty, airtight, plastic flaps. Definitely can't get past those. No way."
-	icon = 'icons/obj/stationobjs.dmi'
+/obj/structure/plasticflaps	//HOW DO YOU CALL THOSE THINGS ANYWAY
+	name = "plastic flaps"
+	desc = "Definitely can't get past those. No way."
+	icon = 'icons/obj/stationobjs.dmi'	//Change this.
 	icon_state = "plasticflaps"
-	armor = list("melee" = 100, "bullet" = 80, "laser" = 80, "energy" = 100, "bomb" = 50, "bio" = 100, "rad" = 100, "fire" = 50, "acid" = 50)
-	density = FALSE
-	anchored = TRUE
-	CanAtmosPass = ATMOS_PASS_NO
-
-/obj/structure/plasticflaps/opaque
-	opacity = TRUE
-
-/obj/structure/plasticflaps/Initialize()
-	. = ..()
-	alpha = 0
-	SSvis_overlays.add_vis_overlay(src, icon, icon_state, ABOVE_MOB_LAYER, plane, dir, add_appearance_flags = RESET_ALPHA) //you see mobs under it, but you hit them like they are above it
-
-/obj/structure/plasticflaps/examine(mob/user)
-	. = ..()
-	if(anchored)
-		. += "<span class='notice'>[src] are <b>screwed</b> to the floor.</span>"
-	else
-		. += "<span class='notice'>[src] are no longer <i>screwed</i> to the floor, and the flaps can be <b>cut</b> apart.</span>"
-
-/obj/structure/plasticflaps/screwdriver_act(mob/living/user, obj/item/W)
-	if(..())
-		return TRUE
-	add_fingerprint(user)
-	var/action = anchored ? "unscrews [src] from" : "screws [src] to"
-	var/uraction = anchored ? "unscrew [src] from " : "screw [src] to"
-	user.visible_message("<span class='warning'>[user] [action] the floor.</span>", "<span class='notice'>You start to [uraction] the floor...</span>", "You hear rustling noises.")
-	if(W.use_tool(src, user, 100, volume=100, extra_checks = CALLBACK(src, .proc/check_anchored_state, anchored)))
-		setAnchored(!anchored)
-		to_chat(user, "<span class='notice'> You [anchored ? "unscrew" : "screw"] [src] from the floor.</span>")
-		return TRUE
-	else
-		return TRUE
-
-/obj/structure/plasticflaps/wirecutter_act(mob/living/user, obj/item/W)
-	if(!anchored)
-		user.visible_message("<span class='warning'>[user] cuts apart [src].</span>", "<span class='notice'>You start to cut apart [src].</span>", "You hear cutting.")
-		if(W.use_tool(src, user, 50, volume=100))
-			if(anchored)
-				return TRUE
-			to_chat(user, "<span class='notice'>You cut apart [src].</span>")
-			var/obj/item/stack/sheet/plastic/five/P = new(loc)
-			P.add_fingerprint(user)
-			qdel(src)
-			return TRUE
-		else
-			return TRUE
-
-/obj/structure/plasticflaps/proc/check_anchored_state(check_anchored)
-	if(anchored != check_anchored)
-		return FALSE
-	return TRUE
-
-/obj/structure/plasticflaps/CanAStarPass(ID, to_dir, caller)
-	if(isliving(caller))
-		if(isbot(caller))
-			return 1
-
-		var/mob/living/M = caller
-		if(!M.ventcrawler && M.mob_size != MOB_SIZE_TINY)
-			return 0
-	var/atom/movable/M = caller
-	if(M && M.pulling)
-		return CanAStarPass(ID, to_dir, M.pulling)
-	return 1 //diseases, stings, etc can pass
+	density = 0
+	anchored = 1
+	layer = 4
 
 /obj/structure/plasticflaps/CanPass(atom/movable/A, turf/T)
-	if(istype(A) && (A.pass_flags & PASSGLASS))
+	if(istype(A) && A.checkpass(PASSGLASS))
 		return prob(60)
 
 	var/obj/structure/bed/B = A
-	if(istype(A, /obj/structure/bed) && (B.has_buckled_mobs() || B.density))//if it's a bed/chair and is dense or someone is buckled, it will not pass
+	if (istype(A, /obj/structure/bed) && (B.buckled_mob || B.density))//if it's a bed/chair and is dense or someone is buckled, it will not pass
 		return 0
 
-	if(istype(A, /obj/structure/closet/cardboard))
+	if (istype(A, /obj/structure/closet/cardboard))
 		var/obj/structure/closet/cardboard/C = A
 		if(C.move_delay)
 			return 0
 
-	if(ismecha(A))
-		return 0
-
-	else if(isliving(A)) // You Shall Not Pass!
+	else if(istype(A, /mob/living)) // You Shall Not Pass!
 		var/mob/living/M = A
-		if(isbot(A)) //Bots understand the secrets
+		if(istype(A,/mob/living/simple_animal/bot/mulebot)) //mulebots can pass
 			return 1
 		if(M.buckled && istype(M.buckled, /mob/living/simple_animal/bot/mulebot)) // mulebot passenger gets a free pass.
 			return 1
@@ -95,17 +30,35 @@
 			return 0
 	return ..()
 
-/obj/structure/plasticflaps/deconstruct(disassembled = TRUE)
-	if(!(flags_1 & NODECONSTRUCT_1))
-		new /obj/item/stack/sheet/plastic/five(loc)
-	qdel(src)
+/obj/structure/plasticflaps/attackby(obj/item/weapon/W, mob/user, params)
+	user.changeNext_move(CLICK_CD_MELEE)
+	..()
 
-/obj/structure/plasticflaps/Initialize()
-	. = ..()
-	air_update_turf(TRUE)
+/obj/structure/plasticflaps/ex_act(severity)
+	..()
+	switch(severity)
+		if (1)
+			qdel(src)
+		if (2)
+			if (prob(50))
+				qdel(src)
+		if (3)
+			if (prob(5))
+				qdel(src)
 
-/obj/structure/plasticflaps/Destroy()
-	var/atom/oldloc = loc
-	. = ..()
-	if (oldloc)
-		oldloc.air_update_turf(1)
+/obj/structure/plasticflaps/mining //A specific type for mining that doesn't allow airflow because of them damn crates
+	name = "airtight plastic flaps"
+	desc = "Heavy duty, airtight, plastic flaps."
+
+/obj/structure/plasticflaps/mining/New() //set the turf below the flaps to block air
+	var/turf/T = get_turf(loc)
+	if(T)
+		T.blocks_air = 1
+	..()
+
+/obj/structure/plasticflaps/mining/Destroy() //lazy hack to set the turf to allow air to pass if it's a simulated floor //wow this is terrible
+	var/turf/T = get_turf(loc)
+	if(T)
+		if(istype(T, /turf/simulated/floor))
+			T.blocks_air = 0
+	return ..()
