+++ code/modules/awaymissions/mission_code/Academy.dm
@@ -1,65 +1,34 @@
-
 //Academy Areas
 
 /area/awaymission/academy
-	name = "Academy Asteroids"
+	name = "\improper Academy Asteroids"
 	icon_state = "away"
 
 /area/awaymission/academy/headmaster
-	name = "Academy Fore Block"
+	name = "\improper Academy Fore Block"
 	icon_state = "away1"
 
 /area/awaymission/academy/classrooms
-	name = "Academy Classroom Block"
+	name = "\improper Academy Classroom Block"
 	icon_state = "away2"
 
 /area/awaymission/academy/academyaft
-	name = "Academy Ship Aft Block"
+	name = "\improper Academy Ship Aft Block"
 	icon_state = "away3"
 
 /area/awaymission/academy/academygate
-	name = "Academy Gateway"
+	name = "\improper Academy Gateway"
 	icon_state = "away4"
 
 /area/awaymission/academy/academycellar
-	name = "Academy Cellar"
+	name = "\improper Academy Cellar"
 	icon_state = "away4"
 
 /area/awaymission/academy/academyengine
-	name = "Academy Engine"
+	name = "\improper Academy Engine"
 	icon_state = "away4"
 
 //Academy Items
-
-/obj/item/paper/fluff/awaymissions/academy/console_maint
-	name = "Console Maintenance"
-	info = "We're upgrading to the latest mainframes for our consoles, the shipment should be in before spring break is over!"
-
-/obj/item/paper/fluff/awaymissions/academy/class/automotive
-	name = "Automotive Repair 101"
-
-/obj/item/paper/fluff/awaymissions/academy/class/pyromancy
-	name = "Pyromancy 250"
-
-/obj/item/paper/fluff/awaymissions/academy/class/biology
-	name = "Biology Lab"
-
-/obj/item/paper/fluff/awaymissions/academy/grade/aplus
-	name = "Summoning Midterm Exam"
-	info = "Grade: A+ Educator's Notes: Excellent form."
-
-/obj/item/paper/fluff/awaymissions/academy/grade/bminus
-	name = "Summoning Midterm Exam"
-	info = "Grade: B- Educator's Notes: Keep applying yourself, you're showing improvement."
-
-/obj/item/paper/fluff/awaymissions/academy/grade/dminus
-	name = "Summoning Midterm Exam"
-	info = "Grade: D- Educator's Notes: SEE ME AFTER CLASS."
-
-/obj/item/paper/fluff/awaymissions/academy/grade/failure
-	name = "Pyromancy Evaluation"
-	info = "Current Grade: F. Educator's Notes: No improvement shown despite multiple private lessons.  Suggest additional tutelage."
-
 
 /obj/singularity/academy
 	dissipate = 0
@@ -84,29 +53,25 @@
 
 /obj/structure/academy_wizard_spawner
 	name = "Academy Defensive System"
-	desc = "Made by Abjuration, Inc."
+	desc = "Made by Abjuration Inc"
 	icon = 'icons/obj/cult.dmi'
 	icon_state = "forge"
-	anchored = TRUE
-	max_integrity = 200
+	anchored = 1
+	var/health = 200
 	var/mob/living/current_wizard = null
 	var/next_check = 0
 	var/cooldown = 600
-	var/faction = ROLE_WIZARD
+	var/faction = "wizard"
+	var/broken = 0
 	var/braindead_check = 0
 
 /obj/structure/academy_wizard_spawner/New()
-	START_PROCESSING(SSobj, src)
-
-/obj/structure/academy_wizard_spawner/Destroy()
-	if(!broken)
-		STOP_PROCESSING(SSobj, src)
-	return ..()
+	SSobj.processing |= src
 
 /obj/structure/academy_wizard_spawner/process()
 	if(next_check < world.time)
 		if(!current_wizard)
-			for(var/mob/living/L in GLOB.player_list)
+			for(var/mob/living/L in player_list)
 				if(L.z == src.z && L.stat != DEAD && !(faction in L.faction))
 					summon_wizard()
 					break
@@ -123,36 +88,67 @@
 		next_check = world.time + cooldown
 
 /obj/structure/academy_wizard_spawner/proc/give_control()
-	set waitfor = FALSE
-
 	if(!current_wizard)
 		return
-	var/list/mob/dead/observer/candidates = pollCandidatesForMob("Do you want to play as Wizard Academy Defender?", ROLE_WIZARD, null, ROLE_WIZARD, 50, current_wizard)
+	spawn(0)
+		var/list/mob/dead/observer/candidates = pollCandidates("Do you want to play as Wizard Academy Defender?", "wizard")
+		var/mob/dead/observer/chosen = null
 
-	if(LAZYLEN(candidates))
-		var/mob/dead/observer/C = pick(candidates)
-		message_admins("[ADMIN_LOOKUPFLW(C)] was spawned as Wizard Academy Defender")
-		current_wizard.ghostize() // on the off chance braindead defender gets back in
-		C.transfer_ckey(current_wizard, FALSE)
+		if(candidates.len)
+			chosen = pick(candidates)
+			message_admins("[key_name_admin(chosen)] was spawned as Wizard Academy Defender")
+			current_wizard.ghostize() // on the off chance braindead defender gets back in
+			current_wizard.key = chosen.key
 
 /obj/structure/academy_wizard_spawner/proc/summon_wizard()
 	var/turf/T = src.loc
+	
 	var/mob/living/carbon/human/wizbody = new(T)
-	wizbody.fully_replace_character_name(wizbody.real_name, "Academy Teacher")
-	wizbody.mind_initialize()
-	var/datum/mind/wizmind = wizbody.mind
+	wizbody.equipOutfit(/datum/outfit/wizard/academy)
+	var/obj/item/weapon/implant/exile/Implant = new/obj/item/weapon/implant/exile(wizbody)
+	Implant.implant(wizbody)
+	wizbody.faction |= "wizard"
+	wizbody.real_name = "Academy Teacher"
+	wizbody.name = "Academy Teacher"
+	
+	var/datum/mind/wizmind = new /datum/mind()
+	wizmind.name = "Wizard Defender"
 	wizmind.special_role = "Academy Defender"
-	wizmind.add_antag_datum(/datum/antagonist/wizard/academy)
-	current_wizard = wizbody
+	var/datum/objective/O = new("Protect Wizard Academy from the intruders")
+	wizmind.objectives += O
+	wizmind.transfer_to(wizbody)
+	ticker.mode.wizards |= wizmind
+	
+	wizmind.AddSpell(new /obj/effect/proc_holder/spell/targeted/ethereal_jaunt)
+	wizmind.AddSpell(new /obj/effect/proc_holder/spell/targeted/projectile/magic_missile)
+	wizmind.AddSpell(new /obj/effect/proc_holder/spell/dumbfire/fireball)
 
+	current_wizard = wizbody
+	
 	give_control()
 
-/obj/structure/academy_wizard_spawner/deconstruct(disassembled = TRUE)
-	if(!broken)
-		broken = 1
+/obj/structure/academy_wizard_spawner/proc/update_status()
+	if(health<0)
 		visible_message("<span class='warning'>[src] breaks down!</span>")
 		icon_state = "forge_off"
-		STOP_PROCESSING(SSobj, src)
+		SSobj.processing.Remove(src)
+		broken = 1
+
+/obj/structure/academy_wizard_spawner/attackby(obj/item/weapon/W, mob/living/user, params)
+	add_fingerprint(user)
+	user.changeNext_move(CLICK_CD_MELEE)
+	if(!broken)
+		health -= W.force
+		update_status()
+	..()
+
+/obj/structure/academy_wizard_spawner/bullet_act(obj/item/projectile/Proj)
+	if(!broken)
+		if((Proj.damage_type == BRUTE || Proj.damage_type == BURN))
+			health -= Proj.damage
+			update_status()
+	..()
+	return
 
 /datum/outfit/wizard/academy
 	name = "Academy Wizard"
@@ -160,40 +156,35 @@
 	r_hand = null
 	suit = /obj/item/clothing/suit/wizrobe/red
 	head = /obj/item/clothing/head/wizard/red
-	backpack_contents = list(/obj/item/storage/box/survival = 1)
+	backpack_contents = list(/obj/item/weapon/storage/box/survival = 1)
 
-/obj/item/dice/d20/fate
-	name = "Die of Fate"
+/obj/item/weapon/dice/d20/fate
+	name = "Dice of Fate"
 	desc = "A die with twenty sides. You can feel unearthly energies radiating from it. Using this might be VERY risky."
 	icon_state = "d20"
 	sides = 20
-	can_be_rigged = FALSE
-	var/reusable = 1
 	var/used = 0
+	var/rigged = -1
 
-/obj/item/dice/d20/fate/one_use
-	reusable = 0
-
-/obj/item/dice/d20/fate/diceroll(mob/user)
+/obj/item/weapon/dice/d20/fate/diceroll(mob/user)
 	..()
 	if(!used)
-		if(!ishuman(user) || !user.mind || (user.mind in SSticker.mode.wizards))
-			to_chat(user, "<span class='warning'>You feel the magic of the dice is restricted to ordinary humans!</span>")
+		if(!ishuman(user) || !user.mind || (user.mind in ticker.mode.wizards))
+			user << "<span class='warning'>You feel the magic of the dice is restricted to ordinary humans!</span>"
 			return
-		if(rigged)
+		if(rigged > 0)
 			effect(user,rigged)
 		else
 			effect(user,result)
 
-/obj/item/dice/d20/fate/equipped(mob/user, slot)
-	if(!ishuman(user) || !user.mind || (user.mind in SSticker.mode.wizards))
-		to_chat(user, "<span class='warning'>You feel the magic of the dice is restricted to ordinary humans! You should leave it alone.</span>")
-		user.dropItemToGround(src)
+/obj/item/weapon/dice/d20/fate/equipped(mob/user, slot)
+	if(!ishuman(user) || !user.mind || (user.mind in ticker.mode.wizards))
+		user << "<span class='warning'>You feel the magic of the dice is restricted to ordinary humans! You should leave it alone.</span>"
+		user.drop_item()
 
 
-/obj/item/dice/d20/fate/proc/effect(var/mob/living/carbon/human/user,roll)
-	if(!reusable)
-		used = 1
+/obj/item/weapon/dice/d20/fate/proc/effect(var/mob/living/carbon/human/user,roll)
+	used = 1
 	visible_message("<span class='userdanger'>The die flare briefly.</span>")
 	switch(roll)
 		if(1)
@@ -204,12 +195,14 @@
 			user.death()
 		if(3)
 			//Swarm of creatures
-			for(var/direction in GLOB.alldirs)
+			for(var/direction in alldirs)
 				var/turf/T = get_turf(src)
-				new /mob/living/simple_animal/hostile/netherworld(get_step(T,direction))
+				new /mob/living/simple_animal/hostile/creature(get_step(T,direction))
 		if(4)
 			//Destroy Equipment
 			for (var/obj/item/I in user)
+				if (istype(I, /obj/item/weapon/implant))
+					continue
 				qdel(I)
 		if(5)
 			//Monkeying
@@ -220,60 +213,61 @@
 			S.speedmod += 1
 		if(7)
 			//Throw
-			user.Stun(60)
+			user.Stun(3)
 			user.adjustBruteLoss(50)
-			var/throw_dir = pick(GLOB.cardinals)
+			var/throw_dir = pick(cardinal)
 			var/atom/throw_target = get_edge_target_turf(user, throw_dir)
 			user.throw_at(throw_target, 200, 4)
 		if(8)
 			//Fueltank Explosion
-			explosion(loc,-1,0,2, flame_range = 2)
+			explosion(src.loc,-1,0,2, flame_range = 2)
 		if(9)
 			//Cold
-			var/datum/disease/D = new /datum/disease/cold()
-			user.ForceContractDisease(D, FALSE, TRUE)
+			var/datum/disease/D = new /datum/disease/cold
+			user.ForceContractDisease(D)
 		if(10)
 			//Nothing
 			visible_message("<span class='notice'>[src] roll perfectly.</span>")
 		if(11)
 			//Cookie
-			var/obj/item/reagent_containers/food/snacks/cookie/C = new(drop_location())
+			var/obj/item/weapon/reagent_containers/food/snacks/cookie/C = new(get_turf(src))
 			C.name = "Cookie of Fate"
 		if(12)
 			//Healing
-			user.revive(full_heal = 1, admin_revive = 1)
+			user.revive()
 		if(13)
 			//Mad Dosh
 			var/turf/Start = get_turf(src)
-			for(var/direction in GLOB.alldirs)
+			for(var/direction in alldirs)
 				var/turf/T = get_step(Start,direction)
 				if(rand(0,1))
 					new /obj/item/stack/spacecash/c1000(T)
 				else
-					var/obj/item/storage/bag/money/M = new(T)
+					var/obj/item/weapon/moneybag/M = new(T)
 					for(var/i in 1 to rand(5,50))
-						new /obj/item/coin/gold(M)
+						new /obj/item/weapon/coin/gold(M)
 		if(14)
 			//Free Gun
-			new /obj/item/gun/ballistic/revolver/mateba(drop_location())
+			new /obj/item/weapon/gun/projectile/revolver/mateba(get_turf(src))
 		if(15)
 			//Random One-use spellbook
-			new /obj/item/book/granter/spell/random(drop_location())
+			new /obj/item/weapon/spellbook/oneuse/random(get_turf(src))
 		if(16)
 			//Servant & Servant Summon
-			var/mob/living/carbon/human/H = new(drop_location())
+			var/mob/living/carbon/human/H = new(get_turf(src))
 			H.equipOutfit(/datum/outfit/butler)
 			var/datum/mind/servant_mind = new /datum/mind()
-			var/datum/antagonist/magic_servant/A = new
-			servant_mind.add_antag_datum(A)
-			A.setup_master(user)
+			var/datum/objective/O = new("Serve [user.real_name].")
+			servant_mind.objectives += O
 			servant_mind.transfer_to(H)
 
-			var/list/mob/dead/observer/candidates = pollCandidatesForMob("Do you want to play as [user.real_name] Servant?", ROLE_WIZARD, null, ROLE_WIZARD, 50, H)
-			if(LAZYLEN(candidates))
-				var/mob/dead/observer/C = pick(candidates)
-				message_admins("[ADMIN_LOOKUPFLW(C)] was spawned as Dice Servant")
-				C.transfer_ckey(H, FALSE)
+			var/list/mob/dead/observer/candidates = pollCandidates("Do you want to play as [user.real_name] Servant?", "wizard")
+			var/mob/dead/observer/chosen = null
+
+			if(candidates.len)
+				chosen = pick(candidates)
+				message_admins("[key_name_admin(chosen)] was spawned as Dice Servant")
+				H.key = chosen.key
 
 			var/obj/effect/proc_holder/spell/targeted/summonmob/S = new
 			S.target_mob = H
@@ -281,13 +275,13 @@
 
 		if(17)
 			//Tator Kit
-			new /obj/item/storage/box/syndicate(drop_location())
+			new /obj/item/weapon/storage/box/syndicate/(get_turf(src))
 		if(18)
 			//Captain ID
-			new /obj/item/card/id/captains_spare(drop_location())
+			new /obj/item/weapon/card/id/captains_spare(get_turf(src))
 		if(19)
 			//Instrinct Resistance
-			to_chat(user, "<span class='notice'>You feel robust.</span>")
+			user << "<span class='notice'>You feel robust.</span>"
 			var/datum/species/S = user.dna.species
 			S.brutemod *= 0.5
 			S.burnmod *= 0.5
@@ -325,13 +319,13 @@
 	if(!target_mob)
 		return
 	var/turf/Start = get_turf(user)
-	for(var/direction in GLOB.alldirs)
+	for(var/direction in alldirs)
 		var/turf/T = get_step(Start,direction)
 		if(!T.density)
 			target_mob.Move(T)
 
 /obj/structure/ladder/unbreakable/rune
-	name = "\improper Teleportation Rune"
+	name = "Teleportation Rune"
 	desc = "Could lead anywhere."
 	icon = 'icons/obj/rune.dmi'
 	icon_state = "1"
@@ -343,6 +337,7 @@
 /obj/structure/ladder/unbreakable/rune/show_fluff_message(up,mob/user)
 	user.visible_message("[user] activates \the [src].","<span class='notice'>You activate \the [src].</span>")
 
-/obj/structure/ladder/unbreakable/rune/use(mob/user, is_ghost=FALSE)
-	if(is_ghost || !(user.mind in SSticker.mode.wizards))
-		..()
+/obj/structure/ladder/can_use(mob/user)
+	if(user.mind in ticker.mode.wizards)
+		return 0
+	return 1
